Order - date form submission works
The orders form lets you select the month/year of the orders you want to
show. My form submitted to the same page URL and kept returning a blank,
white screen. I had to change the form METHOD to GET to make it work. As
a default, form METHOD is POST.